requirements.txt
pyautogui
opencv-python
pytesseract
Pillow
requests
